Extract ActivitySection edit form into ActivityEditForm component
- Add ActivityEditForm component; form state initialises from activity prop, key prop on the parent resets state on selection change (replaces useEffect)
- Remove inline form from ActivitySection, now delegates to ActivityEditForm
- Export Section type from AdminSidebar and import it in page.tsx
